pyproject.toml
[tool.poetry]
name = "hw27-django"
version = "0.1.0"
description = ""
authors = ["macbook <mypur86@yahoo.com>"]
readme = "README.md"
packages = [{include = "hw27_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dateutil = "^2.8.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
pytest = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
